internal: move IsPow
Move pow() from package btf to package internal.
