+++ btf/types_test.go
@@ -38,29 +38,6 @@ func TestSizeof(t *testing.T) {
 	}
 }
 
-func TestPow(t *testing.T) {
-	tests := []struct {
-		n int
-		r bool
-	}{
-		{0, false},
-		{1, true},
-		{2, true},
-		{3, false},
-		{4, true},
-		{5, false},
-		{8, true},
-	}
-
-	for _, tt := range tests {
-		t.Run(fmt.Sprintf("%d", tt.n), func(t *testing.T) {
-			if want, got := tt.r, pow(tt.n); want != got {
-				t.Errorf("unexpected result for n %d; want: %v, got: %v", tt.n, want, got)
-			}
-		})
-	}
-}
-
 func TestCopy(t *testing.T) {
 	_ = Copy((*Void)(nil))
