+++ internal/math.go
@@ -6,3 +6,8 @@ import "golang.org/x/exp/constraints"
 func Align[I constraints.Integer](n, alignment I) I {
 	return (n + alignment - 1) / alignment * alignment
 }
+
+// IsPow returns true if n is a power of two.
+func IsPow[I constraints.Integer](n I) bool {
+	return n != 0 && (n&(n-1)) == 0
+}
